pyproject.toml
[build-system]
requires = ["setuptools>=72"]
build-backend = "setuptools.build_meta"

[project]
name = "pharma-help"
version = "0.1.0"
description = "Context poisoning and indirect prompt injection measurement in agentic AI systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.2",
    "chromadb>=0.5.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-chroma>=0.1.0",
    "pypdf>=4.0.0",
    "httpx>=0.27.0",
    "pytest>=9.0.2",
    "chainlit>=2.0",
    "fastmcp>=2.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "sse-starlette>=2.0.0",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.0.0",
    "langgraph-prebuilt>=1.0.8,<=1.0.8",
]

[project.scripts]
mcp-server      = "pharma_help.mcp.servers.main_server:main"
mcp-fake        = "pharma_help.mcp.servers.fake_server:main"
mcp-confusion   = "pharma_help.mcp.servers.confusion_server:main"
mcp-run-attack  = "pharma_help.mcp.runner.attack_runner:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
